pyproject.toml
[tool.poetry]
name = "absensi-py"
version = "0.1.0"
description = ""
authors = ["addthoriq1304 <thoriq@qti.co.id>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.112.2"}
sqlalchemy = "^2.0.31"
ruff = "^0.5.4"
alembic = "^1.13.2"
psycopg2cffi = "^2.9.0"
asyncpg = "^0.29.0"
bcrypt = "^4.2.0"
pytz = "^2024.1"
pyjwt = "^2.9.0"
typer = "^0.12.3"
polars = "^1.20.0"
xlsxwriter = "^3.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.1"
faker = "^26.3.0"
factory-boy = "^3.3.0"
freezegun = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
